+++ src/test/php/VerifyTest.php
@@ -273,7 +273,8 @@ class VerifyTest extends \PHPUnit_Framework_TestCase
     /**
      * @test
      * @expectedException  PHPUnit_Framework_ExpectationFailedException
-     * @expectedExceptionMessage Parameter 0 for invocation #1 of bovigo\callmap\Verified::aMethod() does not match expected value
+     * @expectedExceptionMessage Parameter $roland at position 0 for invocation #1 of bovigo\callmap\Verified::aMethod() does not match expected value
+     * @since  0.?.?
      */
     public function verifyReceivedPassesExceptionThrownByConstraint()
     {